[refactor] move more ai code into proper location
