+++ functions/ai/openai/whisper.js
@@ -3,7 +3,7 @@
 import OpenAI from "openai";
 import fs from "fs";
 import {logger} from "firebase-functions/v2";
-import {OPENAI_API_KEY} from "../config/config.js";
+import {OPENAI_API_KEY} from "../../config/config.js";
 
 // Initialize the OpenAI client with the API key from environment variables
